🚅 perfs: Simplify psbt util
