+++ src/psbt/bip371.js
@@ -16,8 +16,8 @@ const types_1 = require('../types');
 const transaction_1 = require('../transaction');
 const bip341_1 = require('../payments/bip341');
 const payments_1 = require('../payments');
-const psbtutils_1 = require('./psbtutils');
 const script_1 = require('./input/script');
+const sign_1 = require('./global/sign');
 const toXOnly = pubKey => (pubKey.length === 32 ? pubKey : pubKey.slice(1, 33));
 exports.toXOnly = toXOnly;
 /**
@@ -158,7 +158,7 @@ exports.tapTreeFromList = tapTreeFromList;
 function checkTaprootInputForSigs(input, action) {
   const sigs = extractTaprootSigs(input);
   return sigs.some(sig =>
-    (0, psbtutils_1.signatureBlocksAction)(sig, decodeSchnorrSignature, action),
+    (0, sign_1.signatureBlocksAction)(sig, decodeSchnorrSignature, action),
   );
 }
 exports.checkTaprootInputForSigs = checkTaprootInputForSigs;
