+++ ts_src/psbt/bip371.ts
@@ -20,8 +20,11 @@ import {
 } from '../payments/bip341';
 import { isP2TR, p2tr } from '../payments';
 
-import { signatureBlocksAction } from './psbtutils';
-import { pubkeyPositionInScript, witnessStackToScriptWitness } from './input/script';
+import {
+  pubkeyPositionInScript,
+  witnessStackToScriptWitness,
+} from './input/script';
+import { signatureBlocksAction } from './global/sign';
 
 export const toXOnly = (pubKey: Buffer) =>
   pubKey.length === 32 ? pubKey : pubKey.slice(1, 33);
